// src/common/enums/restaurant-permission.enum.ts
export enum RestaurantPermission {
  MANAGE_BRANDING = 'manage_branding',
  MANAGE_PAYMENTS = 'manage_payments',
  MANAGE_RESTAURANT = 'manage_restaurant',
  MANAGE_STAFF = 'manage_staff',
  MANAGE_MENU = 'manage_menu',
  MANAGE_MARKETING = 'manage_marketing',
  MANAGE_FLOOR = 'manage_floor',
  MANAGE_SCHEDULE = 'manage_schedule',
  MANAGE_ORDER_SETTINGS = 'manage_order_settings',
  MANAGE_BOOKING_SETTINGS = 'manage_booking_settings',
  VIEW_BOOKINGS = 'view_bookings',
  VIEW_ORDERS = 'view_orders',
}